docs: add section to README called "Only for me"
Describe how this repository and any game repository that links to this
repository is only intended to be edited by me.

Also add section called "Terminology" to explain some terms that I used
a lot. One might argue that "submodules" and "pesde" are also referenced
in the README and so should be added as a term, but since these are one-
off occurences, a simple masked link does the job.

Also add "Description" header for the existing section.